Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -1,11 +1,11 @@
 class Solution {
     public int mySqrt(long x) {
 
-        int left = 1;
-        int right = x + 1;
+        long left = 1;
+        long right = x + 1;
 
         while(left < right) {
-            int mid = right + (left - right)/2 ;
+            long mid = right + (left - right)/2 ;
             if(mid*mid > x) {
                 right = mid;
             } else {
